requirements.txt
asgiref==3.7.2
Django==4.2
django-compat==1.0.15
django-environ==0.11.2
django-stubs==4.2.4
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mypy==1.7.1
mypy-extensions==1.0.0
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
tzdata==2023.3
